Noel/requirements.txt
fontawesomefree==6.4.0
django
pymysql
Pillow
django-ckeditor